old real world apps in description
There are various real world applications for our project. The traction control algorithm would provide additional stability and safety in wet weather conditions and would also be useful in racing. The line following algorithm is a simplified version of lane detection and  lane following, which are found in autonomous vehicles and lane keep driver assist systems. The combination of of hardware and software would provide the basic hardware and software needed for an autonomous vehicle to drive on a road with obstacles.